% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{averageScore}
\alias{averageScore}
\title{Average Score}
\usage{
averageScore()
}
\value{
average score (float)
}
\description{
This function computes the average score registerd in the course
"Introduction to Physics"
}
