% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{descriptiveStats}
\alias{descriptiveStats}
\title{Descriptive Statistics}
\usage{
descriptiveStats()
}
\description{
This function provides classical statistics (minimum, maximum,
median, mean, standard deviation), including the skewnees and kurtosis.
The values are reported on the Cullen and Frey graph
}
