% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{gradePlot}
\alias{gradePlot}
\title{Grade Plot}
\usage{
gradePlot()
}
\value{
plotted scores
}
\description{
This function plots an histogram showing scores frequencies, divided by grades
highlited in different colors
}
