% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grades.R
\docType{data}
\name{grades}
\alias{grades}
\title{Grades}
\format{A data frame with 799 rows and 3 variables
\describe{
  \item{count}{progressive number}
  \item{score}{registered test score (0.00-100.00)}
  \item{grade}{grade marked according to the registered score (A-F)}
}}
\usage{
grades
}
\description{
A dataset containing score and grades registered in the "Introduction to
physics" Course.
}
\keyword{datasets}
