% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{gradesEmpDistr}
\alias{gradesEmpDistr}
\title{Grades Empirical Distribution}
\usage{
gradesEmpDistr()
}
\description{
This function plots and characterizes the empirical distribution.
The function return two plots, the left-hand plot is by default the histogram on a density scale
and the right-hand plot the empirical cumulative distribution function
}
