% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{importData}
\alias{importData}
\title{Import Data}
\usage{
importData()
}
\arguments{
\item{none}{does not require parameters}
}
\value{
dataframe containing imported data
}
\description{
This function imports a file from GitHub, which contains physics' grades
}
