% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{medianScore}
\alias{medianScore}
\title{Median Score}
\usage{
medianScore()
}
\value{
median score (float)
}
\description{
This function computes the median score registerd in the
course "Introduction to Physics"
}
