% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{modeScore}
\alias{modeScore}
\title{Mode Score}
\usage{
run the importData function to get a dataframe object to use for this function
}
\arguments{
\item{df}{data frame containing grades' data}
}
\value{
mode score (float)
}
\description{
This function computes the mode score
}
