% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{gradePlot}
\alias{gradePlot}
\title{Grade Plot}
\usage{
run the importData function to get a dataframe object to use for this function
}
\arguments{
\item{df}{data frame containing grads' data}
}
\value{
plotted scores
}
\description{
This function plots an histogram showing scores frequencies, divided by grades
}
