% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsGrades.R
\name{modeScore}
\alias{modeScore}
\title{Mode Score}
\usage{
modeScore()
}
\value{
mode score (float)
}
\description{
This function computes the mode score registerd in the
course "Introduction to Physics"
}
